// clients/socialService/models/SocialFeedPostCreateDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type SocialFeedPostCreateDto = {
    readonly id?: string;
    readonly timestamp?: string;
    title?: string | null;
    message?: string | null;
    socialFeedId?: string | null;
    socialProfileId?: string | null;
};